[refactor] : 상태코드, 구분코드 domain 패키지로 이동
