+++ src/main/java/com/example/stagemate/global/converter/PerformanceTypeConverter.java
@@ -1,6 +1,6 @@
 package com.example.stagemate.global.converter;
 
-import com.example.stagemate.global.code.gubun.PerformanceType;
+import com.example.stagemate.domain.performances.PerformanceType;
 import jakarta.persistence.AttributeConverter;
 import jakarta.persistence.Converter;
